% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voice.R
\name{get_voices}
\alias{get_voices}
\title{Get voices}
\usage{
get_voices()
}
\value{
A tibble with columns voice_id, name, and preview_url.
}
\description{
This function retrieves a list of voices from the URL "https://api.elevenlabs.io/v1/voices"
and returns a tibble containing the voice_id, name, and preview_url for each voice.
}
\examples{
get_voices()

}
\keyword{internal}
