% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voice.R
\name{text_to_speech}
\alias{text_to_speech}
\title{Convert Text to Speech using Eleven Labs API}
\usage{
text_to_speech(
  text,
  api_key = Sys.getenv("ELEVENLABS_API_KEY"),
  voice_name = "Elli",
  output_file = "output.mp3",
  stability = 0,
  similarity_boost = 0,
  ...
)
}
\arguments{
\item{text}{The text to convert to speech}

\item{api_key}{The API key for Eleven Labs API. Default is to use the API key stored in the ELEVENLABS_API_KEY environment variable.}

\item{voice_name}{The name of the voice to use. Default is "Elli".}

\item{output_file}{The name of the output file. Default is "output.mp3"}

\item{stability}{The stability setting, a numeric value between 0 and 1.
Specifies how predictable voice intonation should be. Default is 0}

\item{similarity_boost}{The similarity boost setting, a numeric value between 0 and 1.
Specifies how much the voice should resemble the recordings used to train the voice model. Default is 0}

\item{...}{Additional arguments to be passed on to other methods}
}
\value{
The name of the output file
}
\description{
This function converts text to speech using the Eleven Labs API. The output file is a mp3 format.
}
\examples{

text_to_speech("Hello, how are you today?")

}
